feat!(sql_monitoring): remove tenant-wide dbscan
The idea was to scan the whole tenant for mssql dbs and alerts when there are unmonitored dbs managed by qby.
However, this adds complexity: How to get the permissions for tenant-wide read access?
How to not send the scan events every 5min? Shouldn't such kind of 'we forgot monitoring'-monitoring
be handled for all resources?And isn't that more of an information for our side, rather than
an incident for the customer? For these reasons, I decided to exclude this feature for now.